Implement listpack handling, better errors, ensure that all previously created tests pass
