+++ src/parser/common/mod.rs
@@ -1,4 +1,6 @@
+mod listpack;
 pub mod utils;
 mod ziplist;
 
+pub use listpack::read_list_pack_entry_as_string;
 pub use ziplist::{read_ziplist_entry_string, read_ziplist_metadata};
