+++ src/parser/set.rs
@@ -1,5 +1,7 @@
 use super::common::utils::{read_blob, read_exact, read_length, read_sequence};
-use super::common::{read_ziplist_entry_string, read_ziplist_metadata};
+use super::common::{
+    read_list_pack_entry_as_string, read_ziplist_entry_string, read_ziplist_metadata,
+};
 use crate::types::{RdbError, RdbResult, RdbValue};
 use byteorder::{LittleEndian, ReadBytesExt};
 use std::io::{Cursor, Read};
@@ -103,12 +105,51 @@ pub fn read_sortedset_ziplist<R: Read>(
 
     let last_byte = reader.read_u8()?;
     if last_byte != 0xFF {
-        return Err(RdbError::UnknownEncoding(last_byte));
+        return Err(RdbError::ParsingError {
+            context: "read_sortedset_ziplist",
+            message: format!("Unknown encoding value: {}", last_byte),
+        });
     }
 
     Ok(RdbValue::SortedSet {
         key: key.to_vec(),
         values,
+        expiry,
+    })
+}
+
+pub fn read_set_list_pack<R: Read>(
+    input: &mut R,
+    key: &[u8],
+    expiry: Option<u64>,
+) -> RdbResult<RdbValue> {
+    let listpack = read_blob(input)?;
+    let mut reader = Cursor::new(listpack);
+
+    // Read total bytes and number of elements
+    let total_bytes = reader.read_u32::<LittleEndian>()?;
+    let num_elements = reader.read_u16::<LittleEndian>()?;
+
+    let mut members = Vec::with_capacity(num_elements as usize);
+
+    // Read until we reach the end of the listpack
+    while reader.position() < total_bytes as u64 - 1 {
+        let entry = read_list_pack_entry_as_string(&mut reader)?;
+        members.push(entry);
+    }
+
+    // Verify end byte
+    let last_byte = reader.read_u8()?;
+    if last_byte != 0xFF {
+        return Err(RdbError::ParsingError {
+            context: "read_set_list_pack",
+            message: format!("Unknown encoding value: {}", last_byte),
+        });
+    }
+
+    Ok(RdbValue::Set {
+        key: key.to_vec(),
+        members: members.into_iter().collect(),
         expiry,
     })
 }
