--- tests/mod.rs
@@ -1,2 +1,0 @@
-pub mod format_file_tests;
-pub mod redis_protocol_tests;
